Model the Guess entity
If we want to model the boards, the opponent’s board is the less complex 
of the two as it is nothing but a group of guessed coordinates separated 
into those that hit an island and those that missed.

These lists are modelled using `MapSet`s, as they guarantee that each 
member is unique.