Model the Island entity
More complex than Coordinates and Guesses. They come in five different
shapes. We will also need to determine if an island has been forested or
not—have all the coordinates of the island been planted?

`MapSet`s were chosen again, as
[`MapSet.equal?/2`](https://hexdocs.pm/elixir/MapSet.html#equal?/2)
makes it easy to compare even if the “lists” are not sorted.

[`Enum.reduce_while/3`](https://hexdocs.pm/elixir/Enum.html#reduce_while/3)
is also a worthwhile expansion to understanding the Elixir standard
library.

The use of the [`with`](https://hexdocs.pm/elixir/Kernel.SpecialForms.html#with/1)
special form makes for a beautiful example, complementing the rest of
the code.

The way the runtime check in
```elixir
def new(type, %Coordinate{} = upper_left), do: #…
```
is described helps further understand the benefits and it also helps in
contrasting with the compile-time checks that @specs and dialyzer
augment the code with; @mikestok had mentioned something similar a while
back.