Make the Group class iterable — learn from Jamil
When setting up the `Matrix` class to be iterable, after-the-fact
prototype manipulation was used to add the `Symbol.iterator` method to
the `Matrix` class. The intent was to keep the individual pieces of code
small and self-contained.

As mentioned in the book, there is no need to split the code into small
pieces in a regular program, like the `Group` class. Therefore, declare
the `Symbol.iterator` method directly in the `Group` class.